[#9] cpu.Microcode: Move microcode to subdirectory
