+++ mos6502.cabal
@@ -101,8 +101,10 @@ library
     Cpu.Cpu
     Cpu.Alu
     Cpu.CpuState
-    Cpu.Microcode
     Cpu.Executor
+    Cpu.Microcode.Data
+    Cpu.Microcode.Gen
+    Cpu.Microcode.Map
   default-language: Haskell2010
 
 executable clash
