+++ src/Cpu/Microcode/Map.hs
@@ -1,0 +1,8 @@
+module Cpu.Microcode.Map where
+
+import Clash.Prelude
+import Cpu.Instructions
+import Cpu.Microcode.Data
+
+nextMicrocodeIndex :: (Instruction, AddressingMode) -> MicroOpIndex
+nextMicrocodeIndex = errorX "Not implemented"
